Refactor MainSection in component  and container
