+++ src/containers/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Header from '../components/Header';
-import MainSection from '../components/MainSection';
+import MainSection from '../containers/MainSection';
 
 
 const App = () => (
